Move source into qp folder, create python package and fix imports [make test working]
