+++ qp/gatearray.py
