+++ qp/qmath.py
