+++ setup.py
@@ -1,0 +1,8 @@
+from setuptools import setup
+
+setup(
+    name="Quantum Programming",
+    version="0.1",
+    description="Quantum Programming Simulator",
+    packages=["qp"],
+)
