+++ tests/test_logic_gates.py
@@ -2,12 +2,11 @@ from textwrap import dedent
 
 from numpy import isclose
 
-from drivers.vector.simulator import run_gate_array
-from qrun import run
+from qp.drivers.vector.simulator import run_gate_array
+from qp.qrun import run
 
 
 def test_or():
-
     table = {
         (0, 0, 0): [1, 0, 0, 0, 0, 0, 0, 0],  # -> 0 0 0
         (0, 1, 0): [0, 0, 0, 1, 0, 0, 0, 0],  # -> 0 1 1
@@ -15,7 +14,7 @@ def test_or():
         (1, 1, 0): [0, 0, 0, 0, 0, 0, 0, 1],  # -> 1 1 1
     }
 
-    for setup, facit in table.iteritems():
+    for setup, facit in table.items():
 
         setup = tuple(["X q%d" % i if v == 1 else "" for i, v in enumerate(setup)])
         program = dedent(
